Reorganise les varaibles et les paramètres
