--- openfisca_tunisia/parameters/prestations/amen_social/allocation_base.yaml
@@ -1,17 +1,0 @@
-description: Montants de l'allocation de base du programme Amen social
-values:
-  2019-01-01:
-    value: 180
-  2024-01-01:
-    value: 240
-metadata:
-  unit: currency
-  reference:
-    2019-01-01:
-      - title: Loi n° 2019-10 du 30 janvier 2019 relative au programme AMEN social - Article 2 - Définition des allocations de base
-      - title: Décret n° 2019-317 du 29 mars 2019 fixant les modalités d'application
-    2024-01-01:
-      title: Arrêté ministériel n° 2023-456 du 15 décembre 2023
-  official_journal_date:
-    2020-01-01: "2019-01-31"
-    2024-01-01: "2024-01-03"
