--- openfisca_tunisia/parameters/prestations/amen_social/index.yaml
@@ -1,7 +1,0 @@
-description: Programe Amen social
-metadata:
-  short_label: Amen social
-  order:
-  - allocation_base
-  - aides_ponctuelles
-  - supplements
