+++ openfisca_tunisia/parameters/prestations/non_contributives/amen_social/index.yaml
@@ -1,0 +1,11 @@
+description: Programe Amen social
+metadata:
+  short_label: Amen social
+  order:
+  - allocation_base
+  - aides_ponctuelles
+  - supplements
+  reference:
+    2019-01-01:
+      - title: Loi organique 2019-10 du 30 janvier 2019 relative au programme AMEN social - Article 2 - Définition des allocations de base
+      - title: Décret 2019-317 du 29 mars 2019 fixant les modalités d'application
