+++ openfisca_tunisia/parameters/prestations/non_contributives/amen_social/supplements/personne_a_charge.yaml
@@ -1,4 +1,4 @@
-description: Supplément par personne à charge
+description: Supplément par enafant à charge
 values:
   2019-01-01:
     value: 10.0
