+++ openfisca_tunisia/parameters/prestations/non_contributives/pnafn/allocation.yaml
@@ -1,0 +1,16 @@
+description: Limite d'âge d'un enfant étudiant pour être compté à charge d'un ménage éligible au programme Amen social
+values:
+  2018-02-01:
+    value: 180
+  2015-01-01:
+    value: 150
+  2014-01-01:
+    value: 115
+  2010-01-01:
+    value: 56.7
+  2000-01-01:
+    value: 36.3
+  1987-01-01:
+    value: 7.7
+metadata:
+  unit: currency
