+++ openfisca_tunisia/parameters/prestations/non_contributives/pnafn/index.yaml
@@ -1,0 +1,2 @@
+description:  Programme national d’aide aux familles nécessiteuses (PNAFN)
+documentation: Créé en 1986, comme mesure d'accompagnement du programme d'ajustement structurel, le programme offre un transfert monétaire permanent aux familles pauvres et démunies.
