+++ openfisca_tunisia/variables/prestations/non_contributives/aide_medicale_gratuite.py
@@ -1,0 +1,15 @@
+"""Aide médicale gratuite"""
+
+
+class amg1(Variable):
+    value_type = bool
+    entity = Menage
+    label = 'Aide médicale gratuite - AMG1'
+    definition_period = MONTH
+
+
+class amg1(Variable):
+    value_type = bool
+    entity = Menage
+    label = 'Programme d’assistance médicale à tarifs réduits - AMG2'
+    definition_period = MONTH
